Add StaslkWalk based stack tracer
It seems to be more flexible than CaptureStackBackTrace(),
let's add it and have a play around.

Also added file name and line number decoding to symbolizer
SymbolizeSymFromAddr. Requires debug information generated
into pdb file in order to have it working correct. This seems
to be a common limitation of any symbolizers on Windows.